refactor: #49 Jwt ignoreUrl 제거 및 Method Security에서 Authorization 하도록 수정
